+++ GlockTheClock/assets/scripts/Client/Events/DeathEvent.lua
@@ -1,0 +1,8 @@
+local ns = require('namespace')
+
+ns.deathEvent = ns.class("deathEvent")
+
+function ns.deathEvent:initialize(player)
+	self.player = player
+	LOG("Firing DeathEvent")
+end
